Genie:  Use intrinsic type in enter_system_call()
The size_t typedef won't be visible after we axe the Process.hh include.

+++ lamp/Genie/Genie/Dispatch/kernel_boundary.cc
@@ -33,7 +33,7 @@ namespace Genie
 	{
 		gCurrentProcess->EnterSystemCall();
 		
-		const size_t space = mac::sys::current_thread_stack_space();
+		const unsigned long space = mac::sys::current_thread_stack_space();
 		
 		// space will be 0 if we're not on a Thread Manager stack
